move getFiles
