+++ modules/summaries.js
@@ -55,24 +55,6 @@ function chunkSummaries(summaries, maxChunkLength) {
 }
 
 /**
- * Takes an array of file objects, each with a path property, and returns an array of file objects,
- * each with a path property and a code property containing the file's contents.
- * @param {FileObject[]} files - An array of file objects, each with a path property.
- * @returns {FileObject[]} - An array of file objects,
- * each with a path property and a code property containing the file's contents.
- */
-function getFiles(files){
-  let retFiles=[]
-  for (const file of files) {
-    const pathToFile = file.path;
-    const fileContent = fs.readFileSync(pathToFile, 'utf8');
-    file.code = fileContent
-    retFiles.push(file)
-  }
-  return retFiles
-}
-
-/**
  * Gets all .ai.txt files (summaries)
  * @param {boolean} test - If true, reads files only in the 'benchmarks' directory.
  * @returns {Promise<string>} A string containing all the summaries concatenated together.
@@ -164,7 +146,6 @@ async function generateAndWriteFileSummary(codeBaseDirectory, filePathRelative, 
 
 module.exports = {
     readAllSummaries,
-    getFiles,
     types,
     getSummaries,
     chunkSummaries,
