+++ ui.js
@@ -1,12 +1,14 @@
 // This file is the UI for the user. It accepts a TASK from the user and uses AI to complete the task. Tasks are related with code.
 const chalk = require('chalk');
-const { getTaskInput } = require('./modules/userInputs');
-const { getSummaries, getFiles, chunkSummaries, maxSummaryTokenCount } = require('./modules/summaries');
-const { saveOutput, logPath, updateFile } = require('./modules/fsOutput');
 const agents = require('./agents');
 const yargs = require('yargs');
 const prompts = require('prompts');
-const {printGitDiff} = require('./modules/gitHelper');
+
+const { getTaskInput } = require('./modules/userInputs');
+const { getSummaries, chunkSummaries, maxSummaryTokenCount } = require('./modules/summaries');
+const { saveOutput, logPath, updateFile } = require('./modules/fsOutput');
+const { printGitDiff } = require('./modules/gitHelper');
+const { getFiles } = require('./modules/fsInput');
 
 /**
 @description Asynchronous function that runs an agent function with given variables.
@@ -142,6 +144,7 @@ async function main(task, test=false) {
     reply = await runAgent(agents.getFiles,task, summaries, interactive);
     relevantFiles = relevantFiles.concat(reply.output.relevantFiles)
   }
+  // Fetch code files the agent has deemed relevant
   const files = getFiles(relevantFiles)
   if (files.length == 0) throw new Error("No relevant files found")
